+++ test/table/test_table_with_image.py
@@ -146,7 +146,7 @@ def test_table_with_qrcode(tmp_path):  # issue 771
     assert_pdf_equal(pdf, HERE / "table_with_qrcode.pdf", tmp_path)
 
 
-def test_table_with_page_break_over_image(tmp_path):
+def test_table_with_page_break_over_image_with_heading(tmp_path):
     pdf = FPDF()
     pdf.add_page()
     pdf.set_font("Times", size=16)
@@ -159,7 +159,27 @@ def test_table_with_page_break_over_image(tmp_path):
                     row.cell(img=datum, img_fill_width=True)
                 else:
                     row.cell(datum)
-    assert_pdf_equal(pdf, HERE / "table_with_page_break_over_image.pdf", tmp_path)
+    assert_pdf_equal(
+        pdf, HERE / "table_with_page_break_over_image_with_heading.pdf", tmp_path
+    )
+
+
+def test_table_with_page_break_over_image_without_heading(tmp_path):
+    pdf = FPDF()
+    pdf.add_page()
+    pdf.set_font("Times", size=16)
+    pdf.y = 220
+    with pdf.table(first_row_as_headings=False) as table:
+        for i, data_row in enumerate(TABLE_DATA[:2]):
+            row = table.row()
+            for j, datum in enumerate(data_row):
+                if j == 2 and i > 0:
+                    row.cell(img=datum, img_fill_width=True)
+                else:
+                    row.cell(datum)
+    assert_pdf_equal(
+        pdf, HERE / "table_with_page_break_over_image_without_heading.pdf", tmp_path
+    )
 
 
 def test_table_with_images_and_links(tmp_path):